Send med cookie fra Authorization header til frontend
